Added fetch functionality
So today the code for fetching has been added together with a package to help with visualizing the data. Currently I set it to a radial bar that shows the percentage the student has spent either attentive or inattentive or asleep as those are the definitions Evans says he has thus far. The current test is with manually entered data in the flutter app tomorrow I will try and build the function that returns the json response I expect to use when storing the data before visualizing to the user.