Properly implemented fetch functionality
Updates to student_report.dart and metric_reports.dart: They both are now updated with proper functions that allow to fetch the data and then also generate the chart based off fetched data. Implemented a snackbar to help with informing users if there was an error with fetching the report or a connection takes long.

Will make it a bit more useful say if the error is that there was a connecion time out then the message is that if it is any other type of error that can be easily explained to the user then also make that possible to be displayed in the snackbar.