Made some updates to UI
Today mostly begun working on the circle shapes we saw at the bottom and also adjusting and improving my code moving from guessing numbers for container heights to using MediaQuery so that the app is responsive to different screens. Major focus is screen sizes not orientations so for now it's a portrait strictly app but realized I had forgotten there are different sizes of screens. So onwards, a lot of refactoring of code to make it easier to adjust will be done.