Added way to store fetched data
With SharedPreferences plugin we are now able to keep fetched data and display it in a bar graph on the Homepage in the recent report container.
Also made some slight adjustments to classes and metric_reports.dart. However, I'm not really liking how the data is being displayed on the Home screen so thinking of how to better represent it that's more appealing to the eye