Modified design of Images screen
Added a FAB so that I can end up having just a singular button at the bottom rather than 2 and so that I give a user the chance to also use the camera to take pictures rather than just picking from the gallery.
Also added a timeout to the connection request made to fetch data on the reports screen. Will test again to find a proper acceptable time.
Next thing is adding the ability to pick multiple images because at the time I can only pick one image at a time and that is a drag.