Minor modifications
Made some alterations to snackbar durations and adding the refresh functionality to the homescreen so after the very first fetch of data once the data is cached once we navigate back to the Home Screen the page gets refreshed and the cached data visualized on the homepage as desired. This might be the final commit for this prototype. :)